minios: fix memory leak: packets have to be freed with lwip (issue #40)
When a network interface is used without netmap, the lwip packet
buffers do not get deallocated and the network driver stops receiving
new packets after it runs out of buffer memory. The current fix
registers a new handler which sets the packet destructor on network
interfaces that do not use netmap.